# Optimize and clean up lock