--- a/src/ast.c
+++ b/src/ast.c
@@ -225,13 +225,13 @@ static void jl_init_ast_ctx(jl_ast_context_t *ast_ctx)
 }
 
 // There should be no GC allocation while holding this lock
-JL_DEFINE_MUTEX(flisp)
+static jl_mutex_t flisp_lock;
 static jl_ast_context_list_t *jl_ast_ctx_using = NULL;
 static jl_ast_context_list_t *jl_ast_ctx_freed = NULL;
 
 static jl_ast_context_t *jl_ast_ctx_enter(void)
 {
-    JL_LOCK_NOGC(flisp);
+    JL_LOCK_NOGC(&flisp_lock);
     jl_ast_context_list_t *node;
     jl_ast_context_t *ctx;
     // First check if the current task is using one of the contexts
@@ -239,7 +239,7 @@ static jl_ast_context_t *jl_ast_ctx_enter(void)
         ctx = jl_ast_context_list_item(node);
         if (ctx->task == jl_current_task) {
             ctx->ref++;
-            JL_UNLOCK_NOGC(flisp);
+            JL_UNLOCK_NOGC(&flisp_lock);
             return ctx;
         }
     }
@@ -251,7 +251,7 @@ static jl_ast_context_t *jl_ast_ctx_enter(void)
         ctx->ref = 1;
         ctx->task = jl_current_task;
         ctx->roots = NULL;
-        JL_UNLOCK_NOGC(flisp);
+        JL_UNLOCK_NOGC(&flisp_lock);
         return ctx;
     }
     // Construct a new one if we can't find any
@@ -261,7 +261,7 @@ static jl_ast_context_t *jl_ast_ctx_enter(void)
     ctx->task = jl_current_task;
     node = &ctx->list;
     jl_ast_context_list_insert(&jl_ast_ctx_using, node);
-    JL_UNLOCK_NOGC(flisp);
+    JL_UNLOCK_NOGC(&flisp_lock);
     jl_init_ast_ctx(ctx);
     return ctx;
 }
@@ -270,12 +270,12 @@ static void jl_ast_ctx_leave(jl_ast_context_t *ctx)
 {
     if (--ctx->ref)
         return;
-    JL_LOCK_NOGC(flisp);
+    JL_LOCK_NOGC(&flisp_lock);
     ctx->task = NULL;
     jl_ast_context_list_t *node = &ctx->list;
     jl_ast_context_list_delete(node);
     jl_ast_context_list_insert(&jl_ast_ctx_freed, node);
-    JL_UNLOCK_NOGC(flisp);
+    JL_UNLOCK_NOGC(&flisp_lock);
 }
 
 void jl_init_frontend(void)