--- a/src/dump.c
+++ b/src/dump.c
@@ -25,7 +25,7 @@
 extern "C" {
 #endif
 
-JL_DEFINE_MUTEX(dump)
+static jl_mutex_t dump_lock;
 
 // TODO: put WeakRefs on the weak_refs list during deserialization
 // TODO: handle finalizers
@@ -1957,7 +1957,7 @@ static void jl_save_system_image_to_stream(ios_t *f)
     jl_gc_collect(1); // full
     jl_gc_collect(0); // incremental (sweep finalizers)
     JL_SIGATOMIC_BEGIN();
-    JL_LOCK(dump); // Might GC
+    JL_LOCK(&dump_lock); // Might GC
     int en = jl_gc_enable(0);
     htable_reset(&backref_table, 250000);
     arraylist_new(&reinit_list, 0);
@@ -2001,7 +2001,7 @@ static void jl_save_system_image_to_stream(ios_t *f)
     arraylist_free(&reinit_list);
 
     jl_gc_enable(en);
-    JL_UNLOCK(dump);
+    JL_UNLOCK(&dump_lock);
     JL_SIGATOMIC_END();
 }
 
@@ -2060,7 +2060,7 @@ JL_DLLEXPORT void jl_preload_sysimg_so(const char *fname)
 static void jl_restore_system_image_from_stream(ios_t *f)
 {
     JL_SIGATOMIC_BEGIN();
-    JL_LOCK(dump); // Might GC
+    JL_LOCK(&dump_lock); // Might GC
     int en = jl_gc_enable(0);
     DUMP_MODES last_mode = mode;
     mode = MODE_SYSTEM_IMAGE;
@@ -2118,7 +2118,7 @@ static void jl_restore_system_image_from_stream(ios_t *f)
     jl_gc_enable(en);
     mode = last_mode;
     jl_update_all_fptrs();
-    JL_UNLOCK(dump);
+    JL_UNLOCK(&dump_lock);
     JL_SIGATOMIC_END();
 }
 
@@ -2159,7 +2159,7 @@ JL_DLLEXPORT void jl_restore_system_image_data(const char *buf, size_t len)
 JL_DLLEXPORT jl_array_t *jl_compress_ast(jl_lambda_info_t *li, jl_array_t *ast)
 {
     JL_SIGATOMIC_BEGIN();
-    JL_LOCK(dump); // Might GC
+    JL_LOCK(&dump_lock); // Might GC
     assert(jl_is_lambda_info(li));
     DUMP_MODES last_mode = mode;
     mode = MODE_AST;
@@ -2187,7 +2187,7 @@ JL_DLLEXPORT jl_array_t *jl_compress_ast(jl_lambda_info_t *li, jl_array_t *ast)
     tree_enclosing_module = last_tem;
     jl_gc_enable(en);
     mode = last_mode;
-    JL_UNLOCK(dump);
+    JL_UNLOCK(&dump_lock);
     JL_SIGATOMIC_END();
     return v;
 }
@@ -2195,7 +2195,7 @@ JL_DLLEXPORT jl_array_t *jl_compress_ast(jl_lambda_info_t *li, jl_array_t *ast)
 JL_DLLEXPORT jl_array_t *jl_uncompress_ast(jl_lambda_info_t *li, jl_array_t *data)
 {
     JL_SIGATOMIC_BEGIN();
-    JL_LOCK(dump); // Might GC
+    JL_LOCK(&dump_lock); // Might GC
     assert(jl_is_lambda_info(li));
     assert(jl_is_array(data));
     DUMP_MODES last_mode = mode;
@@ -2213,7 +2213,7 @@ JL_DLLEXPORT jl_array_t *jl_uncompress_ast(jl_lambda_info_t *li, jl_array_t *dat
     tree_literal_values = NULL;
     tree_enclosing_module = NULL;
     mode = last_mode;
-    JL_UNLOCK(dump);
+    JL_UNLOCK(&dump_lock);
     JL_SIGATOMIC_END();
     return v;
 }
@@ -2232,7 +2232,7 @@ JL_DLLEXPORT int jl_save_incremental(const char *fname, jl_array_t *worklist)
     jl_serialize_dependency_list(&f);
 
     JL_SIGATOMIC_BEGIN();
-    JL_LOCK(dump); // Might GC
+    JL_LOCK(&dump_lock); // Might GC
     arraylist_new(&reinit_list, 0);
     htable_new(&backref_table, 5000);
     ptrhash_put(&backref_table, jl_main_module, (char*)HT_NOTFOUND + 1);
@@ -2257,7 +2257,7 @@ JL_DLLEXPORT int jl_save_incremental(const char *fname, jl_array_t *worklist)
     htable_reset(&backref_table, 0);
     arraylist_free(&reinit_list);
     ios_close(&f);
-    JL_UNLOCK(dump);
+    JL_UNLOCK(&dump_lock);
     JL_SIGATOMIC_END();
 
     if (jl_fs_rename(tmpfname, fname) < 0) {
@@ -2391,7 +2391,7 @@ static jl_array_t *_jl_restore_incremental(ios_t *f)
     size_t deplen = read_uint64(f);
     ios_skip(f, deplen); // skip past the dependency list
     JL_SIGATOMIC_BEGIN();
-    JL_LOCK(dump); // Might GC
+    JL_LOCK(&dump_lock); // Might GC
     arraylist_new(&backref_list, 4000);
     arraylist_push(&backref_list, jl_main_module);
     arraylist_new(&flagref_list, 0);
@@ -2417,7 +2417,7 @@ static jl_array_t *_jl_restore_incremental(ios_t *f)
     arraylist_free(&flagref_list);
     arraylist_free(&backref_list);
     ios_close(f);
-    JL_UNLOCK(dump);
+    JL_UNLOCK(&dump_lock);
     JL_SIGATOMIC_END();
 
     JL_GC_PUSH2(&init_order,&restored);