This PR has two related goals.

1. Minor clean up of the lock code and make it available on non-threading build.

    * Wrap the lock in a `jl_mutex_t` type.

        This also fix the last remaining non-C-like syntax introduced by the threading branch...

    * Make `jl_mutex_*` available on non-threading build too.

        This will be useful for synchronization with the signal thread. The `JL_{UN,}LOCK*` macros
        are still disabled for non-threading build although type checks are added to make sure they
        are used correctly. (Working towards a cheap solution of https://github.com/JuliaLang/julia/issues/2622)

        This add ~100 bytes of global variable to `libjulia.so` which is unlikely to cause any change in
        binary size.

        IIUC it is better if the locks doesn't share a cacheline so we can add padding or alignments to
        `jl_mutex_t` later.

2. Optimize lock performance

    * Avoid calling `uv_thread_self` multiple times.
    * Use a different workaround for GCC bug in `store_release` and use it in unlock.

        This saves a lot of cpu cycles compare to sequential consistent store or compare and exchange.

Timing of `LOCK_NOGC` `UNLOCK_NOGC` pair before,

```jl
julia> @benchmark test2()
================ Benchmark Results ========================
     Time per evaluation: 24.80 ns [24.35 ns, 25.25 ns]
Proportion of time in GC: 0.00% [0.00%, 0.00%]
        Memory allocated: 0.00 bytes
   Number of allocations: 0 allocations
       Number of samples: 5701
   Number of evaluations: 458401
         R² of OLS model: 0.951
 Time spent benchmarking: 3.54 s
```

After,
```jl
julia> @benchmark test2()
================ Benchmark Results ========================
     Time per evaluation: 12.07 ns [11.88 ns, 12.27 ns]
Proportion of time in GC: 0.00% [0.00%, 0.00%]
        Memory allocated: 0.00 bytes
   Number of allocations: 0 allocations
       Number of samples: 6101
   Number of evaluations: 670901
         R² of OLS model: 0.959
 Time spent benchmarking: 3.67 s
```