@yuyichao 2016-04-18
Threading CI passed on [travis](https://travis-ci.org/JuliaLang/julia/builds/123811630) and [appveyor](https://ci.appveyor.com/project/JuliaLang/julia/build/1.0.614). The [normal appveyor](https://ci.appveyor.com/project/JuliaLang/julia/build/1.0.611/job/j1a0qkshb1x4aqia) has a weird `DivideError` in the `numbers` test. I've restarted it...