@yuyichao 2016-04-18
Hmm, I believe the numbers test failure is

```jl
julia> Rational(Int8(0), 3) / Complex(3, 2)
ERROR: DivideError: integer division error
 [inlined code] from ./rational.jl:19
 in +(::Rational{Int64}, ::Rational{Int64}) at ./rational.jl:179
 [inlined code] from ./complex.jl:4
 in inv(::Complex{Rational{Int64}}) at ./complex.jl:118
 in /(::Rational{Int64}, ::Complex{Int64}) at ./rational.jl:191
 in eval(::Module, ::Any) at ./boot.jl:237
```