@StefanKarpinski 2016-04-19
I'm generally in favor of not putting ourselves in the position of waiting for upstream changes to get merged – since that has generally been a rather slow process. However, I'm also increasingly concerned about our divergence from standard libuv.