@tkelman 2016-04-19
Upstream on our fork of libuv is us, really. Rebasing the fork to recent versions is broken on windows still, last time that was tried.