@StefanKarpinski 2016-04-19
I'm aware – just making a general comment that we really need to make a concerted effort to upstream some of the stuff on our libuv branch at some point. Not the most urgent thing in the world.