@kpamnany 2016-04-19
It's been about a year since I opened the PR for upstream libuv (https://github.com/libuv/libuv/pull/280, now https://github.com/libuv/libuv/pull/597), so it's clearly not going very quickly. The point of my comment was that since we _are_ using libuv to abstract platform-dependent functionality, it'd be best to keep all the platform-dependence flags in one place, there.